[book2] Section 7-4

+++ app/controllers/staff/programs_controller.rb
@@ -58,8 +58,12 @@ class Staff::ProgramsController < Staff::Base
 
   def destroy
     program = Program.find(params[:id])
-    program.destroy!
-    flash.notice = "プログラムを削除しました。"
+    if program.deletable?
+      program.destroy!
+      flash.notice = "プログラムを削除しました。"
+    else
+      flash.alert = "このプログラムは削除できません。"
+    end
     redirect_to :staff_programs
   end
 end
